integration tests: move bonds integration tests to a separate file
